hash: name/type cleanups
- define FOR_EACH_LIST_ENTRY_OLD2NEW()
- define FOR_EACH_LIST_ENTRY_NEW2OLD()
- hash functions return size_t (since there is no array_index_t)

+++ programs/pluto/hash_table.h
@@ -24,7 +24,7 @@
 
 struct hash_table {
 	const struct list_info info;
-	unsigned long (*hash)(void *data);
+	size_t (*hash)(void *data);
 	long nr_entries; /* approx? */
 	unsigned long nr_slots;
 	struct list_entry *slots;
